+++ external_libs/qwe_Key/qwe_Key.h
@@ -4,7 +4,7 @@
 #include "libhankel.h"
 
 // sasfit strategy 12
-double qwe_Key(double nu, form_factor_f f, double x, void *f_params, double *output,
-               int n_max_iters, double rtol, double atol);
+double qwe_Key(int nu, form_factor_f f, double x, void *f_params, double *output, int n_max_iters,
+               double rtol, double atol);
 
 #endif // QWE_KEY
