+++ tests/test_hankel_DE_Ooura.c
@@ -15,7 +15,7 @@
 
 #define ARRAY_LEN 25
 
-double nu;
+int nu;
 double z;
 
 form_factor_ctx ctx_spheres;
