+++ tests/test_hankel_transform.c
@@ -15,7 +15,7 @@
 
 #define ARRAY_LEN 19
 
-double nu;
+int nu;
 double z;
 
 form_factor_ctx ctx_spheres;
